Reverse the logic of how the viewBox is used for handling the difference in y-axis direction between SVG-space and font-space. Fixes #2
With this change the SVG contours are now drawn in the top-right quadrant and the SVGs are added to the font without a viewBox parameter. (By not having a viewBox parameter the fonts now display correctly in Firefox)
The SVG artwork can still be easily previewed since fonts2svg.py writes the viewBox parameter in a way that makes the artwork display in the bottom-right quadrant.